+++ app/Http/Controllers/Admin/Planning/ResourcePlanningMonitorController.php
@@ -166,6 +166,14 @@ class ResourcePlanningMonitorController extends Controller
                 return $error;
             }
 
+            $existingCount = ResourceOrderItem::where('orderitem_id', $orderItem->id)->count();
+
+            if (! $replace && $existingCount > 0) {
+                return response()->json([
+                    'message' => 'Dit orderitem is al ingepland. Gebruik de optie "vervang bestaande afspraak" of bewerk de bestaande boeking.',
+                ], 422);
+            }
+
             $booking = DB::transaction(function () use ($request, $orderItem, $replace) {
                 if ($replace) {
                     ResourceOrderItem::where('orderitem_id', $orderItem->id)->delete();
